optimize: optimize FenceHandler for oracle
